requirements.txt
apache-airflow-providers-mysql
apache-airflow[slack]
